SO-2022: Change type of complex map group and map priority to integer
- this is necessary to support values greater than Byte.MAX_VALUE (127)

https://snowowl.atlassian.net/browse/SO-2022

Conflicts:
	snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/index/entry/SnomedRefSetMemberIndexEntry.java
	snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/datastore/index/refset/RefSetMemberMutablePropertyUpdater.java
	snomed/com.b2international.snowowl.snomed.refset.core/src/com/b2international/snowowl/snomed/refset/clone/SnomedRefSetCloner.java

+++ snomed/com.b2international.snowowl.snomed.importer.rf2/src/com/b2international/snowowl/snomed/importer/rf2/csv/ComplexMapTypeRefSetRow.java
@@ -36,26 +36,26 @@ public class ComplexMapTypeRefSetRow extends AssociatingRefSetRow {
 	public static final String PROP_CORRELATION_ID = "correlationId";
 	public static final String PROP_MAP_CATEGORY_ID = "mapCategoryId";
 	
-	private byte mapGroup;
-	private byte mapPriority;
+	private int mapGroup;
+	private int mapPriority;
 	private String mapRule;
 	private String mapAdvice;
 	private String correlationId;
 	private String mapCategoryId;
 	
-	public byte getMapGroup() {
+	public int getMapGroup() {
 		return mapGroup;
 	}
 	
-	public void setMapGroup(final byte mapGroup) {
+	public void setMapGroup(final int mapGroup) {
 		this.mapGroup = mapGroup;
 	}
 	
-	public byte getMapPriority() {
+	public int getMapPriority() {
 		return mapPriority;
 	}
 	
-	public void setMapPriority(final byte mapPriority) {
+	public void setMapPriority(final int mapPriority) {
 		this.mapPriority = mapPriority;
 	}
